requirements.txt
crewai>=0.80
crewai-tools
langchain-groq
PyGitHub
python-dotenv
streamlit